pyproject.toml
[tool.poetry]
name = "khlbot-hackpig520"
version = "0.1.0"
description = ""
authors = ["xiaozhu2007 <lz19986912007@163.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
"khl.py" = "^0.2.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"